+++ src/Data/Text/Zipper.hs
@@ -193,104 +193,34 @@ fromText = flip insert empty
 data Span tag = Span tag Text
   deriving (Show)
 
+-- | Text alignment type
+data TextAlignment =
+  TextAlignment_Left
+  | TextAlignment_Right
+  | TextAlignment_Center
+  deriving (Eq, Show)
+
+-- A map from the index (row) of display line to (fst,snd)
+-- fst: leading empty spaces from left (may be negative) to adjust for alignment
+-- snd: the text offset from the beginning of the document
+-- to the first character of the display line
+type OffsetMapWithAlignment = Map Int (Int, Int)
+
+-- helper type representing a single visual line that may be part of a wrapped logical line
+data WrappedLine = WrappedLine
+  { _wrappedLines_text :: Text
+  , _wrappedLines_hiddenWhitespace :: Bool -- ^ 'True' if this line ends with a deleted whitespace character
+  , _wrappedLines_offset :: Int -- ^ offset from beginning of line
+  }
+  deriving (Eq, Show)
+
 -- | Information about the document as it is displayed (i.e., post-wrapping)
 data DisplayLines tag = DisplayLines
   { _displayLines_spans :: [[Span tag]]
-  , _displayLines_offsetMap :: Map Int Int
-  , _displayLines_cursorY :: Int
+  , _displayLines_offsetMap :: OffsetMapWithAlignment
+  , _displayLines_cursorPos :: (Int, Int) -- cursor position relative to upper left hand corner
   }
   deriving (Show)
-
--- | Given a width and a 'TextZipper', produce a list of display lines
--- (i.e., lines of wrapped text) with special attributes applied to
--- certain segments (e.g., the cursor). Additionally, produce the current
--- y-coordinate of the cursor and a mapping from display line number to text
--- offset
-displayLines
-  :: Int -- ^ Width, used for wrapping
-  -> tag -- ^ Metadata for normal characters
-  -> tag -- ^ Metadata for the cursor
-  -> TextZipper -- ^ The text input contents and cursor state
-  -> DisplayLines tag
-displayLines width tag cursorTag (TextZipper lb b a la) =
-  let linesBefore :: [[Text]] -- The wrapped lines before the cursor line
-      linesBefore = map (wrapWithOffset width 0) $ reverse lb
-      linesAfter :: [[Text]] -- The wrapped lines after the cursor line
-      linesAfter = map (wrapWithOffset width 0) la
-      offsets :: Map Int Int
-      offsets = offsetMap $ mconcat
-        [ linesBefore
-        , [wrapWithOffset width 0 $ b <> a]
-        , linesAfter
-        ]
-      spansBefore = map ((:[]) . Span tag) $ concat linesBefore
-      spansAfter = map ((:[]) . Span tag) $ concat linesAfter
-      -- Separate the spans before the cursor into
-      -- * spans that are on earlier display lines (though on the same logical line), and
-      -- * spans that are on the same display line
-      (spansCurrentBefore, spansCurLineBefore) = fromMaybe ([], []) $
-        initLast $ map ((:[]) . Span tag) (wrapWithOffset width 0 b)
-      -- Calculate the number of columns on the cursor's display line before the cursor
-      curLineOffset = spansWidth spansCurLineBefore
-      -- Check whether the spans on the current display line are long enough that
-      -- the cursor has to go to the next line
-      cursorAfterEOL = curLineOffset == width
-      cursorCharWidth = case T.uncons a of
-        Nothing -> 1
-        Just (c, _) -> charWidth c
-      -- Separate the span after the cursor into
-      -- * spans that are on the same display line, and
-      -- * spans that are on later display lines (though on the same logical line)
-      (spansCurLineAfter, spansCurrentAfter) = fromMaybe ([], []) $
-        headTail $ case T.uncons a of
-          Nothing -> [[Span cursorTag " "]]
-          Just (c, rest) ->
-            let o = if cursorAfterEOL then cursorCharWidth else curLineOffset + cursorCharWidth
-                cursor = Span cursorTag (T.singleton c)
-            in case map ((:[]) . Span tag) (wrapWithOffset width o rest) of
-                  [] -> [[cursor]]
-                  (l:ls) -> (cursor : l) : ls
-  in  DisplayLines
-        { _displayLines_spans = concat
-          [ spansBefore
-          , spansCurrentBefore
-          , if cursorAfterEOL
-              then [ spansCurLineBefore, spansCurLineAfter ]
-              else [ spansCurLineBefore <> spansCurLineAfter ]
-          , spansCurrentAfter
-          , spansAfter
-          ]
-        , _displayLines_offsetMap = offsets
-        , _displayLines_cursorY = sum
-          [ length spansBefore
-          , length spansCurrentBefore
-          , if cursorAfterEOL then 1 else 0
-          ]
-        }
-  where
-    initLast :: [a] -> Maybe ([a], a)
-    initLast = \case
-      [] -> Nothing
-      (x:xs) -> case initLast xs of
-        Nothing -> Just ([], x)
-        Just (ys, y) -> Just (x:ys, y)
-    headTail :: [a] -> Maybe (a, [a])
-    headTail = \case
-      [] -> Nothing
-      x:xs -> Just (x, xs)
-
--- | Wraps a logical line of text to fit within the given width. The first
--- wrapped line is offset by the number of columns provided. Subsequent wrapped
--- lines are not.
-wrapWithOffset
-  :: Int -- ^ Maximum width
-  -> Int -- ^ Offset for first line
-  -> Text -- ^ Text to be wrapped
-  -> [Text]
-wrapWithOffset maxWidth _ _ | maxWidth <= 0 = []
-wrapWithOffset maxWidth n xs =
-  let (firstLine, rest) = splitAtWidth (maxWidth - n) xs
-  in firstLine : (fmap (takeWidth maxWidth) . takeWhile (not . T.null) . iterate (dropWidth maxWidth) $ rest)
 
 -- | Split a 'Text' at the given column index. For example
 --
@@ -338,56 +268,6 @@ dropWidth n = snd . splitAtWidth n
 -- on how vty is configed. Please see vty documentation for details.
 charWidth :: Char -> Int
 charWidth = wcwidth
-
-
--- | For a given set of wrapped logical lines, computes a map
--- from display line index to text offset in the original text.
--- This is used to help determine how interactions with the displayed
--- text map back to the original text.
--- For example, given the document @\"AA\\nBBB\\nCCCCCCCC\\n\"@ wrapped to 5 columns,
--- this function will compute the offset in the original document of each character
--- in column 1:
---
--- >   AA...      (0, 0)
--- >   BBB..      (1, 3)
--- >   CCCCC      (2, 7)  -- (this line wraps to the next row)
--- >   CCC..      (3, 12)
--- >   .....      (4, 16)
-offsetMap
-  :: [[Text]] -- ^ The outer list represents logical lines, and the
-              -- inner list represents the display lines into which
-              -- the logical line has been wrapped
-  -> Map Int Int -- ^ A map from the index (row) of display line to
-                 -- the text offset from the beginning of the document
-                 -- to the first character of the display line
-offsetMap ts = evalState (offsetMap' ts) (0, 0)
-  where
-    offsetMap' xs = fmap Map.unions $ forM xs $ \x -> do
-      maps <- forM x $ \line -> do
-        let l = T.length line
-        (dl, o) <- get
-        put (dl + 1, o + l)
-        return $ Map.singleton dl o
-      (dl, o) <- get
-      put (dl, o + 1)
-      return $ Map.insert dl (o + 1) $ Map.unions maps
-
--- | Move the cursor of the given 'TextZipper' to the logical position indicated
--- by the given display line coordinates, using the provided 'DisplayLines'
--- information.  If the x coordinate is beyond the end of a line, the cursor is
--- moved to the end of the line.
-goToDisplayLinePosition :: Int -> Int -> DisplayLines tag -> TextZipper -> TextZipper
-goToDisplayLinePosition x y dl tz =
-  let offset = Map.lookup y $ _displayLines_offsetMap dl
-  in  case offset of
-        Nothing -> tz
-        Just o ->
-          let
-            moveRight = case drop y $ _displayLines_spans dl of
-                [] -> x
-                (s:_) -> charIndexAt x . stream . mconcat . fmap (\(Span _ t) -> t) $ s
-          in  rightN (o + min moveRight x) $ top tz
-
 
 -- | Get the width of the text in a set of 'Span's, taking into account unicode character widths
 spansWidth :: [Span tag] -> Int
@@ -474,33 +354,6 @@ splitWordsAtDisplayWidth maxWidth wwws = reverse $ loop wwws 0 [] where
           else loop (x:xs) 0 [] <> modifyOutForNewLine out
       else loop xs newWidth $ appendOut out x False
 
-data TextAlignment =
-  TextAlignment_Left
-  | TextAlignment_Right
-  | TextAlignment_Center
-  deriving (Eq, Show)
-
--- A map from the index (row) of display line to
--- fst: leading empty spaces from left (may be negative) to adjust for alignment
--- snd: the text offset from the beginning of the document
--- to the first character of the display line
-type OffsetMapWithAlignment = Map Int (Int, Int)
-
--- | Information about the document as it is displayed (i.e., post-wrapping)
-data DisplayLinesWithAlignment tag = DisplayLinesWithAlignment
-    { _displayLinesWithAlignment_spans     :: [[Span tag]]
-    , _displayLinesWithAlignment_offsetMap :: OffsetMapWithAlignment
-    -- cursor position relative to upper left hand corner
-    , _displayLinesWithAlignment_cursorPos :: (Int, Int) -- cursor position relative to upper left hand corner
-    }
-    deriving (Show)
-
-data WrappedLine = WrappedLine
-  { _wrappedLines_text :: Text
-  , _wrappedLines_hiddenWhitespace :: Bool -- ^ 'True' if this line ends with a deleted whitespace character
-  , _wrappedLines_offset :: Int -- ^ offset from beginning of line
-  }
-  deriving (Eq, Show)
 
 
 -- | Wraps a logical line of text to fit within the given width. The first
@@ -555,12 +408,12 @@ offsetMapWithAlignment ts = evalState (offsetMap' ts) (0, 0)
 -- y-coordinate of the cursor and a mapping from display line number to text
 -- offset
 displayLinesWithAlignment
-  :: (Show tag) => TextAlignment
+  :: TextAlignment
   -> Int -- ^ Width, used for wrapping
   -> tag -- ^ Metadata for normal characters
   -> tag -- ^ Metadata for the cursor
   -> TextZipper -- ^ The text input contents and cursor state
-  -> DisplayLinesWithAlignment tag
+  -> DisplayLines tag
 displayLinesWithAlignment alignment width tag cursorTag (TextZipper lb b a la) =
   let linesBefore :: [[WrappedLine]] -- The wrapped lines before the cursor line
       linesBefore = map (wrapWithOffsetAndAlignment alignment width 0) $ reverse lb
@@ -614,8 +467,8 @@ displayLinesWithAlignment alignment width tag cursorTag (TextZipper lb b a la) =
       -- a little silly to convert back to text but whatever, it works
       cursorX = if cursorAfterEOL then 0 else textWidth (mconcat $ fmap (\(Span _ t) -> t) spansCurLineBefore)
 
-  in  DisplayLinesWithAlignment
-        { _displayLinesWithAlignment_spans = concat
+  in  DisplayLines
+        { _displayLines_spans = concat
           [ spansBefore
           , spansCurrentBefore
           , if cursorAfterEOL
@@ -624,8 +477,8 @@ displayLinesWithAlignment alignment width tag cursorTag (TextZipper lb b a la) =
           , spansCurrentAfter
           , spansAfter
           ]
-        , _displayLinesWithAlignment_offsetMap = offsets
-        , _displayLinesWithAlignment_cursorPos = (cursorX, cursorY)
+        , _displayLines_offsetMap = offsets
+        , _displayLines_cursorPos = (cursorX, cursorY)
         }
   where
     initLast :: [a] -> Maybe ([a], a)
@@ -644,15 +497,38 @@ displayLinesWithAlignment alignment width tag cursorTag (TextZipper lb b a la) =
 -- by the given display line coordinates, using the provided 'DisplayLinesWithAlignment'
 -- information.  If the x coordinate is beyond the end of a line, the cursor is
 -- moved to the end of the line.
-goToDisplayLineWithAlignmentPosition :: Int -> Int -> DisplayLinesWithAlignment tag -> TextZipper -> TextZipper
-goToDisplayLineWithAlignmentPosition x y dl tz =
-  let offset = Map.lookup y $ _displayLinesWithAlignment_offsetMap dl
+goToDisplayLinePosition :: Int -> Int -> DisplayLines tag -> TextZipper -> TextZipper
+goToDisplayLinePosition x y dl tz =
+  let offset = Map.lookup y $ _displayLines_offsetMap dl
   in  case offset of
         Nothing -> tz
         Just (alignOff,o) ->
           let
             trueX = max 0 (x - alignOff)
-            moveRight = case drop y $ _displayLinesWithAlignment_spans dl of
+            moveRight = case drop y $ _displayLines_spans dl of
                 []    -> 0
                 (s:_) -> charIndexAt trueX . stream . mconcat . fmap (\(Span _ t) -> t) $ s
           in  rightN (o + moveRight) $ top tz
+
+-- | Given a width and a 'TextZipper', produce a list of display lines
+-- (i.e., lines of wrapped text) with special attributes applied to
+-- certain segments (e.g., the cursor). Additionally, produce the current
+-- y-coordinate of the cursor and a mapping from display line number to text
+-- offset
+displayLines
+  :: Int -- ^ Width, used for wrapping
+  -> tag -- ^ Metadata for normal characters
+  -> tag -- ^ Metadata for the cursor
+  -> TextZipper -- ^ The text input contents and cursor state
+  -> DisplayLines tag
+displayLines = displayLinesWithAlignment TextAlignment_Left
+
+-- | Wraps a logical line of text to fit within the given width. The first
+-- wrapped line is offset by the number of columns provided. Subsequent wrapped
+-- lines are not.
+wrapWithOffset
+  :: Int -- ^ Maximum width
+  -> Int -- ^ Offset for first line
+  -> Text -- ^ Text to be wrapped
+  -> [Text]
+wrapWithOffset maxWidth n xs = fmap _wrappedLines_text $ wrapWithOffsetAndAlignment TextAlignment_Left maxWidth n xs